Refactor tests file
